requirements.txt
asgiref==3.5.0
Django==4.0.2
django-phonenumber-field==6.1.0
djangorestframework==3.13.1
phonenumbers==8.12.43
psycopg2==2.9.3
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
